Split project into multiple modules
